Replace inline code with snapshots
