+++ spock-specs/src/main/resources/org/spockframework/smoke/ast/AstSpec__astToSourceFeatureBody_can_render_everything__Groovy_4_0_2__.groovy
@@ -1,0 +1,14 @@
+package apackage
+
+import spock.lang.*
+
+@org.spockframework.runtime.model.SpecMetadata(filename = 'script.groovy', line = 1)
+public class apackage.ASpec extends spock.lang.Specification implements groovy.lang.GroovyObject {
+
+    @org.spockframework.runtime.model.FeatureMetadata(name = 'a feature', ordinal = 0, line = 1, blocks = [@org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.SETUP, texts = [])], parameterNames = [])
+    public void $spock_feature_0_0() {
+        java.lang.Object nothing = null
+        this.getSpecificationContext().getMockController().leaveScope()
+    }
+
+}
