+++ spock-specs/src/main/resources/org/spockframework/smoke/ast/AstSpec__astToSourceFeatureBody_shows_compile_error_in_source.groovy
@@ -1,0 +1,7 @@
+Unable to produce AST for this phase due to earlier compilation error:
+startup failed:
+script.groovy: 3: 'expect' is not allowed here; instead, use one of: [and, then] @ line 3, column 13.
+       expect: 'none'
+               ^
+
+1 error
