+++ spock-specs/src/main/resources/org/spockframework/smoke/ast/AstSpec__groovy_3_language_features.groovy
@@ -1,0 +1,26 @@
+public void loop() {
+    do {
+        this.println('once')
+    } while (false)
+}
+
+public void methodRef() {
+    [].forEach(java.lang.System::'print')
+}
+
+public void lambdas() {
+    java.lang.Object lambda = ( java.lang.Object x) -> {
+        x * x }
+    java.lang.Object lambdaMultiArg = ( int a, int b) -> {
+        a <=> b }
+    java.lang.Object lambdaNoArg = ( ) -> {
+        throw new java.lang.RuntimeException('bam')
+    }
+}
+
+public void blockStatements() {
+    with_label:
+    {
+        this.println(foo)
+    }
+}
