+++ spock-specs/src/test/groovy/org/spockframework/runtime/ExtensionClassesLoaderSpec.groovy
@@ -3,6 +3,7 @@ package org.spockframework.runtime
 import org.spockframework.report.log.ReportLogConfiguration
 
 import org.spockframework.runtime.extension.builtin.*
+import org.spockframework.specs.extension.SnapshotConfig
 import org.spockframework.tempdir.TempDirConfiguration
 import spock.config.RunnerConfiguration
 import spock.lang.Specification
@@ -22,6 +23,6 @@ class ExtensionClassesLoaderSpec extends Specification {
     def result = new ExtensionClassesLoader().loadConfigClassesFromDefaultLocation()
 
     then:
-    result == [RunnerConfiguration, ReportLogConfiguration, TempDirConfiguration]
+    result == [SnapshotConfig, RunnerConfiguration, ReportLogConfiguration, TempDirConfiguration]
   }
 }
