+++ spock-specs/src/test/groovy/org/spockframework/smoke/ast/condition/CollectionConditionAstSpec.groovy
@@ -1,10 +1,17 @@
 package org.spockframework.smoke.ast.condition
 
 import org.spockframework.EmbeddedSpecification
+import org.spockframework.specs.extension.Snapshot
+import org.spockframework.specs.extension.Snapshotter
 
 class CollectionConditionAstSpec extends EmbeddedSpecification {
+  @Snapshot(extension = 'groovy')
+  Snapshotter snapshotter
 
   def "collection condition matchCollectionsAsSet is transformed correctly"() {
+    given:
+    snapshotter.featureBody()
+
     when:
     def result = compiler.transpileFeatureBody('''
         given:
@@ -14,25 +21,13 @@ class CollectionConditionAstSpec extends EmbeddedSpecification {
     ''')
 
     then:
-    /* language=Groovy */
-    result.source == '''\
-@org.spockframework.runtime.model.FeatureMetadata(name = 'a feature', ordinal = 0, line = 1, blocks = [@org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.SETUP, texts = []), @org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.EXPECT, texts = [])], parameterNames = [])
-public void $spock_feature_0_0() {
-    org.spockframework.runtime.ErrorCollector $spock_errorCollector = org.spockframework.runtime.ErrorRethrower.INSTANCE
-    org.spockframework.runtime.ValueRecorder $spock_valueRecorder = new org.spockframework.runtime.ValueRecorder()
-    java.lang.Object x = [1]
-    try {
-        org.spockframework.runtime.SpockRuntime.verifyMethodCondition($spock_errorCollector, $spock_valueRecorder.reset(), 'x =~ [1]', 4, 9, null, org.spockframework.runtime.SpockRuntime, $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(0), 'matchCollectionsAsSet'), new java.lang.Object[]{$spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(1), x), $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(3), [$spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(2), 1)])}, $spock_valueRecorder.realizeNas(6, false), false, 5)
-    }
-    catch (java.lang.Throwable $spock_condition_throwable) {
-        org.spockframework.runtime.SpockRuntime.conditionFailedWithException($spock_errorCollector, $spock_valueRecorder, 'x =~ [1]', 4, 9, null, $spock_condition_throwable)}
-    finally {
-    }
-    this.getSpecificationContext().getMockController().leaveScope()
-}'''
+    snapshotter.assertThat(result.source).matchesSnapshot()
   }
 
   def "collection condition matchCollectionsInAnyOrder is transformed correctly"() {
+    given:
+    snapshotter.featureBody()
+
     when:
     def result = compiler.transpileFeatureBody('''
         given:
@@ -42,25 +37,13 @@ public void $spock_feature_0_0() {
     ''')
 
     then:
-    /* language=Groovy */
-    result.source == '''\
-@org.spockframework.runtime.model.FeatureMetadata(name = 'a feature', ordinal = 0, line = 1, blocks = [@org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.SETUP, texts = []), @org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.EXPECT, texts = [])], parameterNames = [])
-public void $spock_feature_0_0() {
-    org.spockframework.runtime.ErrorCollector $spock_errorCollector = org.spockframework.runtime.ErrorRethrower.INSTANCE
-    org.spockframework.runtime.ValueRecorder $spock_valueRecorder = new org.spockframework.runtime.ValueRecorder()
-    java.lang.Object x = [1]
-    try {
-        org.spockframework.runtime.SpockRuntime.verifyMethodCondition($spock_errorCollector, $spock_valueRecorder.reset(), 'x ==~ [1]', 4, 9, null, org.spockframework.runtime.SpockRuntime, $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(0), 'matchCollectionsInAnyOrder'), new java.lang.Object[]{$spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(1), x), $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(3), [$spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(2), 1)])}, $spock_valueRecorder.realizeNas(6, false), false, 5)
-    }
-    catch (java.lang.Throwable $spock_condition_throwable) {
-        org.spockframework.runtime.SpockRuntime.conditionFailedWithException($spock_errorCollector, $spock_valueRecorder, 'x ==~ [1]', 4, 9, null, $spock_condition_throwable)}
-    finally {
-    }
-    this.getSpecificationContext().getMockController().leaveScope()
-}'''
+    snapshotter.assertThat(result.source).matchesSnapshot()
   }
 
   def "regex find conditions are transformed correctly"() {
+    given:
+    snapshotter.featureBody()
+
     when:
     def result = compiler.transpileFeatureBody('''
         given:
@@ -70,25 +53,13 @@ public void $spock_feature_0_0() {
     ''')
 
     then:
-    /* language=Groovy */
-    result.source == '''\
-@org.spockframework.runtime.model.FeatureMetadata(name = 'a feature', ordinal = 0, line = 1, blocks = [@org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.SETUP, texts = []), @org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.EXPECT, texts = [])], parameterNames = [])
-public void $spock_feature_0_0() {
-    org.spockframework.runtime.ErrorCollector $spock_errorCollector = org.spockframework.runtime.ErrorRethrower.INSTANCE
-    org.spockframework.runtime.ValueRecorder $spock_valueRecorder = new org.spockframework.runtime.ValueRecorder()
-    java.lang.Object x = '[1]\'
-    try {
-        org.spockframework.runtime.SpockRuntime.verifyCondition($spock_errorCollector, $spock_valueRecorder.reset(), 'x =~ /\\\\d/', 4, 9, null, $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(2), $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(0), x) =~ $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(1), '\\\\d')))
-    }
-    catch (java.lang.Throwable $spock_condition_throwable) {
-        org.spockframework.runtime.SpockRuntime.conditionFailedWithException($spock_errorCollector, $spock_valueRecorder, 'x =~ /\\\\d/', 4, 9, null, $spock_condition_throwable)}
-    finally {
-    }
-    this.getSpecificationContext().getMockController().leaveScope()
-}'''
+    snapshotter.assertThat(result.source).matchesSnapshot()
   }
 
   def "regex match conditions are transformed correctly"() {
+    given:
+    snapshotter.featureBody()
+
     when:
     def result = compiler.transpileFeatureBody('''
         given:
@@ -98,21 +69,6 @@ public void $spock_feature_0_0() {
     ''')
 
     then:
-    /* language=Groovy */
-    result.source == '''\
-@org.spockframework.runtime.model.FeatureMetadata(name = 'a feature', ordinal = 0, line = 1, blocks = [@org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.SETUP, texts = []), @org.spockframework.runtime.model.BlockMetadata(kind = org.spockframework.runtime.model.BlockKind.EXPECT, texts = [])], parameterNames = [])
-public void $spock_feature_0_0() {
-    org.spockframework.runtime.ErrorCollector $spock_errorCollector = org.spockframework.runtime.ErrorRethrower.INSTANCE
-    org.spockframework.runtime.ValueRecorder $spock_valueRecorder = new org.spockframework.runtime.ValueRecorder()
-    java.lang.Object x = 'a1b'
-    try {
-        org.spockframework.runtime.SpockRuntime.verifyCondition($spock_errorCollector, $spock_valueRecorder.reset(), 'x ==~ /a\\\\db/', 4, 9, null, $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(2), $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(0), x) ==~ $spock_valueRecorder.record($spock_valueRecorder.startRecordingValue(1), 'a\\\\db')))
-    }
-    catch (java.lang.Throwable $spock_condition_throwable) {
-        org.spockframework.runtime.SpockRuntime.conditionFailedWithException($spock_errorCollector, $spock_valueRecorder, 'x ==~ /a\\\\db/', 4, 9, null, $spock_condition_throwable)}
-    finally {
-    }
-    this.getSpecificationContext().getMockController().leaveScope()
-}'''
+    snapshotter.assertThat(result.source).matchesSnapshot()
   }
 }
