+++ spock-specs/src/test/groovy/org/spockframework/smoke/ast/condition/ExceptionConditionsAstSpec.groovy
@@ -1,13 +1,20 @@
 package org.spockframework.smoke.ast.condition
 
 import org.spockframework.EmbeddedSpecification
+import org.spockframework.specs.extension.Snapshot
+import org.spockframework.specs.extension.Snapshotter
 import spock.lang.Issue
 import spock.util.Show
 
 class ExceptionConditionsAstSpec extends EmbeddedSpecification {
+  @Snapshot(extension = 'groovy')
+  Snapshotter snapshotter
 
   @Issue("https://github.com/spockframework/spock/issues/1266")
   def "thrown rewrite keeps correct method reference"() {
+    given:
+    snapshotter.specBody()
+
     when:
     def result = compiler.transpileSpecBody('''
 def "cleanup blocks don't destroy method reference when invocation is assigned to variable with the same name"() {
@@ -21,31 +28,16 @@ def "cleanup blocks don't destroy method reference when invocation is assigned t
 def foobar() {
   throw new IllegalStateException("foo")
 }''', EnumSet.of(Show.METHODS))
+
     then:
-    result.source == '''\
-public java.lang.Object foobar() {
-    throw new java.lang.IllegalStateException('foo')
-}
-
-public void $spock_feature_0_0() {
-    java.lang.Object foobar
-    this.getSpecificationContext().setThrownException(null)
-    try {
-        foobar = this.foobar()
-    }
-    catch (java.lang.Throwable $spock_ex) {
-        this.getSpecificationContext().setThrownException($spock_ex)
-    }
-    finally {
-    }
-    this.thrownImpl(null, null, java.lang.IllegalStateException)
-    this.getSpecificationContext().getMockController().leaveScope()
-}'''
-
+    snapshotter.assertThat(result.source).matchesSnapshot()
   }
 
   @Issue("https://github.com/spockframework/spock/issues/1332")
   def "thrown rewrite keeps correct method reference for multi-assignments"() {
+    given:
+    snapshotter.specBody()
+
     when:
     def result = compiler.transpileSpecBody('''
 def "cleanup blocks don't destroy method reference when invocation is assigned to variable with the same name"() {
@@ -59,26 +51,8 @@ def "cleanup blocks don't destroy method reference when invocation is assigned t
 def foobar() {
   throw new IllegalStateException("foo")
 }''', EnumSet.of(Show.METHODS))
+
     then:
-    result.source == '''\
-public java.lang.Object foobar() {
-    throw new java.lang.IllegalStateException('foo')
-}
-
-public void $spock_feature_0_0() {
-    def (java.lang.Object foobar, java.lang.Object b) = [null, null]
-    this.getSpecificationContext().setThrownException(null)
-    try {
-        (foobar, b) = this.foobar()
-    }
-    catch (java.lang.Throwable $spock_ex) {
-        this.getSpecificationContext().setThrownException($spock_ex)
-    }
-    finally {
-    }
-    this.thrownImpl(null, null, java.lang.IllegalStateException)
-    this.getSpecificationContext().getMockController().leaveScope()
-}'''
-
+    snapshotter.assertThat(result.source).matchesSnapshot()
   }
 }
